pyproject.toml
[project]
name = "mcpx"
version = "1.0.0"
description = "MCPX HTTP MCP Server - A production-ready MCP server with utility tools"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "MCPX Contributors"}
]
keywords = ["mcp", "model-context-protocol", "fastmcp", "tools", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastmcp>=2.13.0.2",
    "httpx>=0.28.1",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
mcpx = "mcp_server.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
